% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print_dots}
\alias{print_dots}
\title{Print the ... argument}
\usage{
print_dots(...)
}
\arguments{
  \item{...}{The ... argument from the function to be
  analysed}
}
\description{
A really simple function that explores what happens when use ...
}

